feat: Restructured project files into a new folder hierarchy
