--- Articles_of_Association_sample.txt
@@ -1,14 +1,0 @@
-Article 1: Jurisdiction
-This agreement shall be governed by the laws of the UAE Federal Courts.
-
-Article 2: Signature
-Signed by the authorized signatory.
-
-Article 3: Confidentiality
-The parties may share confidential information, but no obligation shall arise.
-
-Article 4: Governing Law
-This agreement is subject to the governing law of the jurisdiction.
-
-Article 5: Dispute Resolution
-Any dispute shall be resolved according to ADGM regulations.
