--- checklist.json
@@ -1,9 +1,0 @@
-{
-  "Company Incorporation": [
-    "Articles of Association",
-    "Memorandum of Association",
-    "Board Resolution",
-    "UBO Declaration",
-    "Register of Members and Directors"
-  ]
-}
