+++ data/checklists/adgm_letters_and_permits.txt
