+++ data/checklists/adgm_registration_incorporation.txt
