+++ data/checklists/adgm_setting_up.txt
