+++ docs/official_references/Templates_SHReso_AmendmentArticles-v1-20220107.docx
