--- llm_local.py
@@ -1,9 +1,0 @@
-import subprocess
-
-def query_ollama(prompt, model="mistral"):
-    result = subprocess.run(
-        ["ollama", "run", model],
-        input=prompt.encode(),
-        stdout=subprocess.PIPE
-    )
-    return result.stdout.decode()
