+++ src/checklist/checklist.json
@@ -1,0 +1,9 @@
+{
+  "Company Incorporation": [
+    "Articles of Association",
+    "Memorandum of Association",
+    "Board Resolution",
+    "UBO Declaration",
+    "Register of Members and Directors"
+  ]
+}
