--- summary.json
@@ -1,30 +1,0 @@
-{
-  "process": "Company Incorporation",
-  "documents_uploaded": 1,
-  "required_documents": 5,
-  "missing_documents": [
-    "Articles of Association",
-    "Memorandum of Association",
-    "UBO Declaration",
-    "Register of Members and Directors"
-  ],
-  "issues_found": [
-    {
-      "document": "Board Resolution",
-      "issues": [
-        {
-          "issue": "Incorrect jurisdiction reference",
-          "suggestion": "Update to ADGM Courts",
-          "severity": "High",
-          "source": "ADGM Regulations 2020"
-        },
-        {
-          "issue": "Missing signatory section",
-          "suggestion": "Add signatory section",
-          "severity": "Medium",
-          "source": "ADGM Document Templates"
-        }
-      ]
-    }
-  ]
-}
